added exception for Tag

+++ service/src/main/java/com/epam/esm/service/GiftCertificateService.java
@@ -51,6 +51,10 @@ public class GiftCertificateService {
     }
 
     public void delete(int id) {
+        boolean existById = giftCertificatesDao.isExistById(id);
+        if(!existById){
+            throw new CertificateNotFoundException(id);
+        }
         giftCertificatesDao.delete(id);
     }
 
@@ -78,10 +82,7 @@ public class GiftCertificateService {
         if (giftCertificate.getTag()!=null) {
             tagVerification.checkAndSaveTagIfNotExist(giftCertificate);
         }
-        boolean existById = giftCertificatesDao.isExistById(giftCertificate.getId());
-        if (!existById){
-            throw new CertificateNotFoundException("name");
-        }
+
         GiftCertificate certificateFromDb = giftCertificatesDao.findById(giftCertificate.getId());
         boolean equalsCertificates = certificateFromDb.equals(giftCertificate);
         if (!equalsCertificates) {
